Corrections due to more stringent compiler warnings.

+++ src/libs/acctlib/treenode.h
@@ -134,11 +134,11 @@ IMPLEMENT_ARCHIVE_LIST(CTreeNodeList);
 // EXISTING_CODE
 //---------------------------------------------------------------------------
 inline unsigned commonPrefix(const SFString& _t, const SFString& _u) {
-    unsigned s = min((unsigned)_t.length(), (unsigned)_u.length());
-    for (unsigned i = 0 ; ; ++i)
+    int32_t s = (int32_t)min(_t.length(), _u.length());
+    for (int32_t i = 0 ; ; ++i)
         if (i == s || (_t[i] != _u[i]))
-            return i;
-    return s;
+            return (unsigned)i;
+    return (unsigned)s;
 }
 // EXISTING_CODE
 }  // namespace qblocks
